Update frame_bb_dist.py
Just changed the hydrogen/ nitrogen. It will give error residue 17 but is the glycine which is residue 19. I forgot how we fixed it. The only way I can think of it was if statements, but I thought we had an efficient way last time probably in your old code. I will work on it after the game.